requirements.txt
bert_score
matplotlib
wmd
py-rouge
nltk
tqdm
six
scikit-learn
pathlib
configargparse
bottle
ujson
GPUtil
coloredlogs
inflect
unidecode
psutil
tensorflow
wandb
torch==1.6.0
transformers==3.5.1
datasets
pytorch-lightning==1.0.4